I2C bus recovery on Duet 2 and accelerometer fixes
TwoWire owns the TWI pin muxing now and recovers a bus whose slave is left holding SDA low, M122 reports the count as bus recoveries
Added M122 P1009 to leave the DueX expander holding SDA low, to test the recovery
Remote accelerometer collections above 65535 samples work
Bug fix: An expansion board restarting during an accelerometer collection left the file open and M956 reporting that the accelerometer was already collecting until M999

+++ src/CAN/CanInterface.cpp
@@ -1733,7 +1733,7 @@ GCodeResult CanInterface::DeleteFilamentMonitor(DriverId driver, GCodeBuffer* gb
 
 # if SUPPORT_ACCELEROMETERS
 
-GCodeResult CanInterface::StartAccelerometer(DriverId device, uint8_t axes, uint16_t numSamples, uint8_t mode, const GCodeBuffer& gb, const StringRef& reply) THROWS(GCodeException)
+GCodeResult CanInterface::StartAccelerometer(DriverId device, uint8_t axes, uint32_t numSamples, uint8_t mode, const GCodeBuffer& gb, const StringRef& reply) THROWS(GCodeException)
 {
 	CanMessageBuffer* const buf = AllocateBuffer(&gb);
 	const CanRequestId rid = CanInterface::AllocateRequestId(device.boardAddress, buf);
